reorganized project structure and added retrofit for api calls
