+++ app/src/main/java/com/isep/series/activities/MainActivity.java
@@ -1,4 +1,4 @@
-package com.isep.series;
+package com.isep.series.activities;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
@@ -7,9 +7,12 @@ import androidx.fragment.app.Fragment;
 import android.os.Bundle;
 import android.view.MenuItem;
 
-import com.google.android.material.bottomnavigation.BottomNavigationItemView;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 import com.google.android.material.navigation.NavigationBarView;
+import com.isep.series.R;
+import com.isep.series.fragments.DiscoverFragment;
+import com.isep.series.fragments.ProfileFragment;
+import com.isep.series.fragments.SeriesFragment;
 
 public class MainActivity extends AppCompatActivity {
