+++ app/src/main/java/com/isep/series/fragments/DiscoverFragment.java
@@ -1,20 +1,21 @@
-package com.isep.series;
+package com.isep.series.fragments;
 
-import android.content.Context;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.view.Window;
 import android.view.WindowManager;
-import android.widget.Adapter;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
-import androidx.viewpager2.widget.ViewPager2;
+
+import com.isep.series.R;
+import com.isep.series.models.Series;
+import com.isep.series.adapters.DiscoverAdapter;
 
 import java.util.ArrayList;
 import java.util.List;
