+++ app/src/main/java/com/isep/series/fragments/ProfileFragment.java
@@ -1,4 +1,4 @@
-package com.isep.series;
+package com.isep.series.fragments;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -10,6 +10,10 @@ import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.isep.series.R;
+import com.isep.series.models.Series;
+import com.isep.series.adapters.ProfileAdapter;
 
 import java.util.ArrayList;
 import java.util.List;
