+++ app/src/main/java/com/isep/series/helpers/API.java
@@ -1,0 +1,13 @@
+package com.isep.series.helpers;
+import retrofit2.Call;
+import retrofit2.http.GET;
+
+
+public interface API {
+
+    //@GET("http://api.exchangeratesapi.io/v1/symbols?access_key=779028d721bbab39cf32a1f511c48c64")
+    //Call<CurrencySymbol> getssymbols();
+
+    //@GET("http://api.exchangeratesapi.io/v1/latest?access_key=779028d721bbab39cf32a1f511c48c64")
+    //Call<CurrencyRates> getCurrencyRates();
+}
