+++ app/src/main/java/com/isep/series/models/Series.java
@@ -1,4 +1,4 @@
-package com.isep.series;
+package com.isep.series.models;
 
 public class Series {
